refactor(backend): extract the vector members into engine members
Moves upsertEmbedding, upsertEmbeddingBatch, deleteEmbedding,
deleteEmbeddingBatch, vectorSearch, deleteSchemaVectorSlotContribution,
createVectorIndex, and dropVectorIndex out of the PostgreSQL and SQLite
operation-backend factories into engine/members/vector-members.ts as
createVectorMembers(deps), spread from both createPostgresOperationBackend
and createSqliteOperationBackend exactly where the inline definitions used
to sit.

The two genuine per-engine differences become explicit deps rather than
being re-spelled in the shared body:

- applySearchOverrides resolves and applies this engine's per-search
  efSearch handling and runs the built query. PostgreSQL validates
  efSearch against pgvector's ceiling, computes any transaction-local GUC
  overrides (SET LOCAL hnsw.*/ivfflat.*), and runs through its
  transaction-wrapped runner; no bundled SQLite engine has a per-search
  frontier knob, so its arm calls the same portable
  resolveEfSearchOverride predicate purely to refuse an override, then
  runs the query unmodified.
- runIndexBuild runs one already-built ANN index-creation statement.
  PostgreSQL routes it through the serial fallback that retries under
  parallel_workers = 0 when the parallel build exhausts dynamic shared
  memory; SQLite just runs the statement.

hybridSearch is not extracted (it composes the vector leg with the
dialect-owned single-statement hybrid SQL) and stays inline in each
factory, now under its own smaller vectorEmbeddingMethods block. The
conditional spread that omits the whole embedding write/search group when
no vector strategy is wired is preserved verbatim via a discriminated
CreateVectorMembersDeps union (vectorStrategy: undefined collapses every
other field), and createVectorIndex/dropVectorIndex/
deleteSchemaVectorSlotContribution keep their unconditional presence with
an internal early return, matching today's shape exactly.

No DRIZZLE_ZONE entry is needed: the new file has no direct Drizzle
import, same as the existing fulltext-members.ts.

Also carries forward the rationale comments the extraction had dropped
rather than moved: createVectorIndex's shared doc comment now explains
why the explicit buildCreateIndex step runs after ensureVectorSlot even
when a strategy's ownedTables already folds index DDL into the table
(picks up the slot's declared index params, and covers slots whose index
intent changed after first materialization); PostgreSQL's runIndexBuild
dep regains the comment on why the concurrent flag is honored (no
write-blocking lock on a live table) and why that requires the
autocommit execRun path. The hybridSearch comments in postgres.ts and
sqlite.ts that pointed at "vectorSearch above" now point at
createVectorMembers's vectorSearch in engine/members/vector-members.ts,
where that member actually lives. EngineLateMembers' doc comment no
longer cites GUC-wrapped vector search as a dialect-owned example, since
vector search is now a shared body in vector-members.ts with only its
runner threaded through as a dep; it cites the raw-SQL escape hatch's
driver-level binding and decoding instead.
